% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_matrix.R
\name{as_coord_matrix}
\alias{as_coord_matrix}
\alias{coord_matrix}
\alias{st_as_sfc.coord_matrix}
\alias{st_as_sf.coord_matrix}
\title{Coerce R Object to Waypoints}
\usage{
as_coord_matrix(x, ...)

st_as_sfc.coord_matrix(x)

st_as_sf.coord_matrix(x)
}
\arguments{
\item{x}{a \code{matrix}, a \code{data.frame}, a \code{sf::sfc_POINT} object, or a
\code{character} string in the form \code{"lon1,lat1;lon2,lat2;..."}.

If \code{x} is a \code{matrix} the columns are assumed to be in the order \verb{lat, lon},
except if the column names \verb{lat, lon} are already present.

If \code{x} is a \code{data.frame} it must contain a column
named either \code{"lat"} or \code{"latitude"} and a column named either \code{"lon"},
\code{"lng"} or \code{"longitude"} (all case insenstive)}
}
\value{
\code{as_coord_matrix()} returns a \code{coord_matrix} object: A numeric \code{matrix}
with the columns \code{lat}and \code{lon} (in that order)

The \code{st_as_sf()} and \code{st_as_sf()} methwods for \code{coord_matrix}
\code{\link[sf:st_point]{sf::st_point()}} objects.
}
\description{
Coerce R Object to Waypoints
}
