% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waypoints.R
\name{as_waypoints}
\alias{as_waypoints}
\alias{waypoints}
\alias{st_as_sfc.waypoints}
\alias{st_as_sf.waypoints}
\title{Coerce R Object to Waypoints for gisco::gc_route}
\usage{
as_waypoints(x)

st_as_sfc.waypoints(x)

st_as_sf.waypoints(x)
}
\arguments{
\item{x}{a \code{matrix}, a \code{data.frame}, a \code{sf::sfc_POINT} object, or a
\code{character} string in the form \code{"lon1,lat1;lon2,lat2;..."}.

If \code{x} is a \code{matrix} the columns are assumed to be in the order \verb{lat, lon},
except if the column names \verb{lat, lon} are already present.

If \code{x} is a \code{data.frame} it must contain a column
named either \code{"lat"} or \code{"latitude"} and a column named either \code{"lon"},
\code{"lng"} or \code{"longitude"} (all case insenstive)}
}
\value{
\code{as_waypoints()} returns a \code{waypoints} object: A numeric \code{matrix}
with the columns \code{lat}and \code{lon} (in that order)

The \code{st_as_sf()} and \code{st_as_sf()} methwods for \code{waypoints}
\code{\link[sf:st_point]{sf::st_point()}} objects.
}
\description{
Create valid inputs for \code{\link[=gc_route]{gc_route()}}
}
\examples{
as_waypoints("16.373831,48.208171;14.437802,50.07554;11.581979,48.135129")
}
\seealso{
\link{waypoints_to_string}
}
