% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickmap.R
\name{convert_colors_to_hex}
\alias{convert_colors_to_hex}
\title{Convert R color names to hex codes}
\usage{
convert_colors_to_hex(color_vector)
}
\arguments{
\item{color_vector}{Character vector of R color names or hex codes}
}
\value{
Character vector of hex codes (uppercase format)
}
\description{
Convert R color names to hex codes
}
\seealso{
Other colour: 
\code{\link{get_contrast_text_color}()},
\code{\link{lighten_color}()}
}
\concept{colour}
