% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickmap.R
\name{get_colour_legend}
\alias{get_colour_legend}
\title{Extract the legend fields from a colour scale}
\usage{
get_colour_legend(scale = "lbrut_no2")
}
\arguments{
\item{scale}{Name of a YAML scale in \verb{inst/config/scales/}}
}
\value{
List of \code{colors}, \code{labels}, \code{title} and \code{thresholds}
}
\description{
Extract the legend fields from a colour scale
}
\seealso{
Other colour: 
\code{\link{assign_colour}()},
\code{\link{convert_colors_to_hex}()},
\code{\link{get_contrast_text_color}()},
\code{\link{lighten_color}()}
}
\concept{colour}
\keyword{internal}
