% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickmap.R
\name{load_colour_scale}
\alias{load_colour_scale}
\title{Load colour scale from YAML configuration file}
\usage{
load_colour_scale(scale_name)
}
\arguments{
\item{scale_name}{Character string, name of the colour scale (e.g., "who_no2")}
}
\value{
List containing colour scale definition
}
\description{
Load colour scale from YAML configuration file
}
\seealso{
Other config: 
\code{\link{load_theme}()},
\code{\link{load_yaml_config}()},
\code{\link{show_borough_colours}()}
}
\concept{config}
