% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickmap.R
\name{load_yaml_config}
\alias{load_yaml_config}
\title{Load data source configuration from YAML file}
\usage{
load_yaml_config(name, subdirectory = NULL, list_available = FALSE)
}
\arguments{
\item{source_name}{Character string, name of the data source (e.g., "dt_sites", "bl_nodes", "schools")}
}
\value{
List containing data source configuration with fields: id, label, icon_shape, static
}
\description{
Load data source configuration from YAML file
}
\seealso{
Other config: 
\code{\link{load_colour_scale}()},
\code{\link{load_theme}()},
\code{\link{show_borough_colours}()}
}
\concept{config}
\keyword{internal}
