% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smartmap-package.R
\docType{package}
\name{smartmap-package}
\alias{smartmap}
\alias{smartmap-package}
\title{smartmap: Smartly Create Maps from R Objects}
\description{
View a wide range of common formats for spatial data as 
  interactive leaflet maps.
}
\author{
\strong{Maintainer}: Stefan Fleck \email{stefan.b.fleck@gmail.com} (\href{https://orcid.org/0000-0003-3344-9851}{ORCID})

}
\keyword{internal}
