% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_matrix.R
\name{st_as_sf.coord_matrix}
\alias{st_as_sf.coord_matrix}
\alias{st_as_sfc.coord_matrix}
\title{Coerce coordinate matrices to sf or sfc}
\usage{
\method{st_as_sf}{coord_matrix}(x, ...)

\method{st_as_sfc}{coord_matrix}(x, ...)
}
\arguments{
\item{x}{a \link{coord_matrix}}

\item{...}{ignored}
}
\value{
The \code{st_as_sf()} and \code{st_as_sf()} methwods for \code{coord_matrix}
\code{\link[sf:st_point]{sf::st_point()}} objects.
}
\description{
Coerce coordinate matrices to sf or sfc
}
