% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_matrix.R
\name{st_as_sf}
\alias{st_as_sf}
\alias{st_as_sf.coord_matrix}
\title{Convert coordinate matrices to sf objects}
\usage{
\method{st_as_sf}{coord_matrix}(x, ...)
}
\arguments{
\item{x}{a \link{coord_matrix}}

\item{...}{ignored}
}
\value{
an \code{\link[sf:sf]{sf::sf()}} object with an \code{sfc_POINT}-geometry column
}
\description{
Convert coordinate matrices to sf objects
}
\seealso{
\code{\link[sf:st_as_sf]{sf::st_as_sf()}}
}
