% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_matrix.R
\name{st_as_sfc}
\alias{st_as_sfc}
\alias{st_as_sfc.coord_matrix}
\title{Convert coordinate matrices to sfc objects}
\usage{
\method{st_as_sfc}{coord_matrix}(x, ...)
}
\arguments{
\item{x}{a \link{coord_matrix}}

\item{...}{ignored}
}
\value{
an \code{\link[sf:sfc]{sf::sfc()}} object of subclass \code{sfc_POINT}
}
\description{
Convert coordinate matrices to sfc objects
}
\seealso{
\code{\link[sf:st_as_sfc]{sf::st_as_sfc()}}
}
