% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind.R
\name{add_wind_layer}
\alias{add_wind_layer}
\title{Attach the wind overlay to an interactive map widget}
\usage{
add_wind_layer(map, wind, display_times, bbox, wind_style = NULL)
}
\description{
Attach the wind overlay to an interactive map widget
}
\keyword{internal}
