% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickmap.R
\name{calculate_max_range_width}
\alias{calculate_max_range_width}
\title{Calculate maximum range width for fixed-width legend blocks}
\usage{
calculate_max_range_width(labels)
}
\arguments{
\item{labels}{Character vector of legend labels}
}
\value{
Integer number of characters for the longest range (including symbol space)
}
\description{
Calculate maximum range width for fixed-width legend blocks
}
\seealso{
Other legend: 
\code{\link{generate_indicator_bar}()},
\code{\link{generate_indicator_html}()},
\code{\link{get_symbol_for_index}()},
\code{\link{parse_legend_label}()},
\code{\link{ramp_position}()},
\code{\link{trim_colour_scale}()}
}
\concept{legend}
