% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm_layer.R
\name{qm_meta}
\alias{qm_meta}
\title{Layer metadata of a qm_layer}
\usage{
qm_meta(x)
}
\arguments{
\item{x}{A qm_layer}
}
\value{
Named list: value_col, time_col, label_col, shape, name, resolution
}
\description{
Layer metadata of a qm_layer
}
\seealso{
Other atomic unit: 
\code{\link{from_csv}()},
\code{\link{from_openair}()},
\code{\link{from_rdata}()},
\code{\link{qm_layer}()}
}
\concept{atomic unit}
