% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmap.R
\name{qmap}
\alias{qmap}
\alias{qmap.leaflet}
\alias{qmap.sf}
\alias{qmap.default}
\alias{qmap.sfc}
\alias{qmap.sfg}
\alias{qmap.matrix}
\alias{qmap.character}
\title{View spatial objects as interactive leaflet maps}
\usage{
qmap(
  x,
  ...,
  tools = TRUE,
  provider = getOption("qmap.providers", "OpenStreetMap")
)

\method{qmap}{leaflet}(
  x,
  ...,
  tools = TRUE,
  provider = getOption("qmap.providers", "OpenStreetMap")
)

\method{qmap}{sf}(
  x,
  ...,
  tools = TRUE,
  provider = getOption("qmap.providers", "OpenStreetMap")
)

\method{qmap}{default}(
  x,
  labels = NULL,
  ...,
  tools = TRUE,
  provider = getOption("qmap.providers", "OpenStreetMap")
)

\method{qmap}{sfc}(
  x,
  ...,
  tools = TRUE,
  provider = getOption("qmap.providers", "OpenStreetMap")
)

\method{qmap}{sfg}(
  x,
  ...,
  tools = TRUE,
  provider = getOption("qmap.providers", "OpenStreetMap")
)

\method{qmap}{matrix}(
  x,
  labels = NULL,
  ...,
  tools = TRUE,
  provider = getOption("qmap.providers", "OpenStreetMap")
)

\method{qmap}{character}(
  x,
  labels = NULL,
  ...,
  tools = TRUE,
  provider = getOption("qmap.providers", "OpenStreetMap")
)
}
\arguments{
\item{x}{quickmap currently supports coercion of the following data types:
\itemize{
\item \code{matrix}: Either a matrix with named \code{latitude} and \code{longitude} columns
or an unnamed two column matrix containing longitude and lattiude
(in that order)
\item a \code{data.frame} mit named \code{latitude} and \code{longitude} columns
\item an \link[sf:sfc_POINT]{sf::sfc_POINT} object
\item a named or unmmaned \code{numeric} vecot of length 2 containign a single
longitude-latitude coordinate pair
}}

\item{...}{passed on to methods.}

\item{tools}{\code{logical} scalar. If \code{TRUE} show additional tools on the
resulting map (such as a ruler and the ability to switch between several
background tiles)}

\item{provider}{\code{character} vector. Name of one or several valid providers
for \code{\link[leaflet:addProviderTiles]{leaflet::addProviderTiles()}}. If \code{tools == TRUE} you will be able
to switch interactively between all supplied providers on the returned
leaflet map, if \code{tools == FALSE} only the first provider will be used.}

\item{labels}{an optional \code{character} vector of popup labels}
}
\value{
a \link[leaflet:leaflet]{leaflet::leaflet} object
}
\description{
Can be used to preview spatial \R objects
}
\note{
\code{quickmap.default()} looks if an \code{\link[sf:st_as_sf]{sf::st_as_sf()}},
\code{\link[sf:st_as_sfc]{sf::st_as_sfc()}} or \code{\link[qmap:as_coord_matrix]{qmap::as_coord_matrix()}} method exists for \code{x} (in
that order). If you are a package developer and want to support quickmap
for a custom S3 class in your package, it is enough to provide one of these
methods.
}
\examples{

wp <- matrix(
  c(16.419684, 48.186065,
    16.373894, 48.207853,
    16.285887, 48.083053)
  byrow = TRUE,
  ncol = 2
)

\donttest{
  qmap(wp)
  qmap(c(16.419684, 48.186065))
}


}
