% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickmap-package.R
\docType{package}
\name{quickmap-package}
\alias{quickmap}
\alias{quickmap-package}
\title{quickmap: Quickly Create Maps from R objects}
\description{
Preview a wide range of common formats for spatial data as 
  interactive leaflet maps.
}
\author{
\strong{Maintainer}: Stefan Fleck \email{stefan.b.fleck@gmail.com} (\href{https://orcid.org/0000-0003-3344-9851}{ORCID})

}
\keyword{internal}
